Add home view and routing so grids can be open in standalone view without tabs.
