+++ src/app/app.component.scss
@@ -1,147 +1,5 @@
-@use "igniteui-angular/theming" as *;
-
-$custom-icon-button-theme: icon-button-theme(
-  $background: #212121,
-  $foreground: #ffffff,
-  $disabled-background: #9e9e9e,
-  $disabled-foreground: #ffffff
-);
-
-// Pass the css-vars to the `css-vars` mixin
-@include css-vars($custom-icon-button-theme);
-
 :host {
   display: block;
   width: 100%;
   height: 100%;
-  padding: 8px;
-
-  ::ng-deep {
-    @include css-vars($custom-icon-button-theme);
-  }
 }
-
-igx-chip {
-  --igx-chip-hover-background: var(--ig-gray-300);
-  --igx-chip-focus-background: var(--ig-gray-300);
-}
-
-.demo-container {
-  width: 100%;
-  height: 100%;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  gap: 24px;
-}
-
-.router-container {
-  width: 100%;
-  height: calc(100% - 208px);
-}
-
-.tab-container {
-  width: 100%;
-  min-height: 132px;
-  display: flex;
-  flex-direction: row;
-  justify-content: space-between;
-}
-
-.tab-item-container {
-  height: 100%;
-  display: inline-flex;
-}
-
-.tab-item {
-  height: 100%;
-  display: flex;
-  flex-grow: 1;
-  flex-basis: 0;
-  flex-direction: column;
-  justify-content: space-between;
-  gap: 8px;
-  padding: 12px;
-  cursor: pointer;
-  user-select: none;
-  border-bottom: 1px solid var(--ig-gray-300);
-
-
-  &--selected {
-    border-bottom: 3px solid var(--ig-primary-500);
-  }
-
-  &:hover {
-    background: hsl(from var(--ig-gray-100) h s l/0.5);
-  }
-}
-
-.tab-header {
-  display: flex;
-  flex-direction: row;
-  gap: 8px;
-  color: var(--ig-gray-900);
-  font-size: 20px;
-  font-weight: 600;
-  line-height: 24px;
-  letter-spacing: 0.15px;
-  --ig-size: var(--ig-size-medium);
-
-  &--disabled {
-    color: var(--ig-gray-700) !important;
-  }
-}
-
-.tab-content {
-  display: flex;
-  flex-direction: column;
-  gap: 4px;
-  color: var(--ig-gray-900);
-  font-size: 14px;
-  font-weight: 400;
-  font-family: "aktiv-grotesk", sans-serif;
-  line-height: 20px;
-  letter-spacing: 0.25px;
-
-  &--disabled {
-    color: var(--ig-gray-700) !important;
-  }
-}
-
-.tab-actions {
-  display: flex;
-  flex-direction: row;
-  align-items: center;
-  justify-content: space-between;
-  --ig-size: var(--ig-size-small);
-
-  a {
-    padding-bottom: 2px;
-    background: linear-gradient(0deg, var(--ig-primary-500), var(--ig-primary-500)) no-repeat right bottom / 0 var(--bg-h);
-    transition: background-size 350ms;
-    --bg-h: 2px;
-
-    &:where(:hover, :focus-visible) {
-      background-size: 100% var(--bg-h);
-      background-position-x: left;
-    }
-  }
-
-  .link--disabled {
-    color: var(--ig-primary-200);
-    background: none !important;
-    background-size: 0% var(--bg-h) !important;
-  }
-
-}
-
-.learn-text {
-  text-decoration: none;
-  color: var(--ig-primary-500);
-  font-size: 14px;
-  font-weight: 700;
-  font-family: "aktiv-grotesk", sans-serif;
-  line-height: 20px;
-}
-
-
